requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
autopep8==2.3.2
boto3==1.37.31
botocore==1.37.31
certifi==2025.1.31
cffi==1.17.1
Django==4.2.19
django-extensions==4.1
django-silk==5.3.2
django-storages==1.14.6
djangorestframework==3.16.0
dotenv==0.9.9
gprof2dot==2025.4.14
graphviz==0.20.3
gunicorn==23.0.0
jmespath==1.0.1
minio==7.2.15
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.22.0

python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==1.26.20
